split client, service and mcp classes into separate files
